requirements.txt
albumentations==1.2.1
efficientnet-pytorch==0.7.1
h5py==3.1.0
hydra-core==1.2.0
importlib-metadata==4.12.0
matplotlib==3.5.2
mediapipe==0.8.4
mpose==1.0.26
opencv-contrib-python==4.6.0.66
opencv-python-headless==4.6.0.66
pandas==1.3.5
Pillow==9.2.0
pyrealsense2==2.50.0.3812
PyYAML==6.0
pyzmq==23.2.0
tensorboard==2.9.1
tensorflow-gpu==2.5.0
timm==0.6.7
tqdm==4.64.0
matplotlib==3.5.2
mediapipe==0.8.4
numpy==1.19.3
opencv-contrib-python==4.6.0.66
opencv-python==4.5.5.62
pyrealsense2==2.50.0.3812
pyzmq==23.2.0